// src/proto/kafkapb/kafka_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
// @generated from file kafkapb/kafka.proto (package dev.reduction.kafka, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { Timestamp } from "@bufbuild/protobuf/wkt";
import { file_google_protobuf_timestamp } from "@bufbuild/protobuf/wkt";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file kafkapb/kafka.proto.
 */
export const file_kafkapb_kafka: GenFile = /*@__PURE__*/
  fileDesc("ChNrYWZrYXBiL2thZmthLnByb3RvEhNkZXYucmVkdWN0aW9uLmthZmthIqMBCgZSZWNvcmQSDQoFdG9waWMYASABKAkSEQoJcGFydGl0aW9uGAIgASgFEgsKA2tleRgDIAEoDBINCgV2YWx1ZRgEIAEoDBItCgl0aW1lc3RhbXAYBSABKAsyGi5nb29nbGUucHJvdG9idWYuVGltZXN0YW1wEiwKB2hlYWRlcnMYBiADKAsyGy5kZXYucmVkdWN0aW9uLmthZmthLkhlYWRlciIkCgZIZWFkZXISCwoDa2V5GAEgASgJEg0KBXZhbHVlGAIgASgMQipaKHJlZHVjdGlvbi5kZXYvcmVkdWN0aW9uLXByb3RvY29sL2thZmthcGJiBnByb3RvMw", [file_google_protobuf_timestamp]);

/**
 * @generated from message dev.reduction.kafka.Record
 */
export type Record = Message<"dev.reduction.kafka.Record"> & {
  /**
   * @generated from field: string topic = 1;
   */
  topic: string;

  /**
   * @generated from field: int32 partition = 2;
   */
  partition: number;

  /**
   * @generated from field: bytes key = 3;
   */
  key: Uint8Array;

  /**
   * @generated from field: bytes value = 4;
   */
  value: Uint8Array;

  /**
   * @generated from field: google.protobuf.Timestamp timestamp = 5;
   */
  timestamp?: Timestamp;

  /**
   * @generated from field: repeated dev.reduction.kafka.Header headers = 6;
   */
  headers: Header[];
};

/**
 * Describes the message dev.reduction.kafka.Record.
 * Use `create(RecordSchema)` to create a new message.
 */
export const RecordSchema: GenMessage<Record> = /*@__PURE__*/
  messageDesc(file_kafkapb_kafka, 0);

/**
 * @generated from message dev.reduction.kafka.Header
 */
export type Header = Message<"dev.reduction.kafka.Header"> & {
  /**
   * @generated from field: string key = 1;
   */
  key: string;

  /**
   * @generated from field: bytes value = 2;
   */
  value: Uint8Array;
};

/**
 * Describes the message dev.reduction.kafka.Header.
 * Use `create(HeaderSchema)` to create a new message.
 */
export const HeaderSchema: GenMessage<Header> = /*@__PURE__*/
  messageDesc(file_kafkapb_kafka, 1);